Make _StencilComp a float
This prevents a warning from being displayed.

+++ Packages/best.crux.buffer-viewer/Shaders/Stencil Match.shader
@@ -3,7 +3,7 @@ Shader "Hidden/chemicalcrux/Buffer Viewer/Stencil Match"
     Properties
     {
         _StencilRef ("Stencil Ref", Integer) = 0
-        _StencilComp ("Stencil Comp", Integer) = 0
+        _StencilComp ("Stencil Comp", Float) = 0
     }
     SubShader
     {
